+++ library/src/main/scala/giter8/construct/utils/MavenHelper.scala
@@ -16,6 +16,7 @@
  */
 
 package giter8
+package construct.utils
 
 import scala.xml.{NodeSeq, XML}
 import org.apache.http.HttpResponse
